+++ opesci/templates/copyright.txt
